requirements.txt
numpy
opencv-python
cryptography
onnxruntime
streamlit
Pillow
